Add initial implementation of <complex.h>
Provide creal*() and cimag*() functions with a basic parameterized test to
ensure that we are correctly handling these components.

Since we don't yet have <stdio.h> or <assert.h> or any other part of the C
standard library implemented, we compute a value and return it as the exit code
from main() to help us validate our implementation.