Test a single function at-a-time in complex_test
Rather than testing an expression consisting of the output of  several
functions, test each function independently, so we can see which function(s) are
working vs. which ones are failing and in which context. This will make it
easier to see what's working vs.  what isn't.

Also added similar tests for cimagf(), cimagl(), crealf(), and creall(), but
these are disabled because they currently fail on Linux (but pass on macOS).