Replace macOS 10.15 with macOS 11 and 12
macOS 10.15 is deprecated:
https://github.com/actions/runner-images#available-images
so we need to replace it with the current macOS 11 and 12.

They have different versions of Xcode installed, so we parameterize the runner
to specify different pairs of (macOS version, Xcode version) for separate jobs.

Additionally, Ubuntu 22.04 is now available, so add it to the matrix, while also
keeping Ubuntu 20.04, which isn't yet deprecated.