Split GitHub Actions workflow into Linux + macOS
This will let us show separate badges in the README to show the status
independently, with more meaningful names rather than the generic "build &
test".

While this may also provide additional parallelism, we don't actually need it:
the Linux build is extremely fast (~6 sec) and the macOS is the slow one.